+++ src/context.rs
@@ -1,36 +1,21 @@
-use std::{env, ffi::OsString, ops};
+use std::{env, ffi::OsString};
 
 use anyhow::{bail, Result};
 use camino::Utf8PathBuf;
 use cargo_metadata::PackageId;
 
-use crate::{
-    cargo::{self, Cargo},
-    cli::Args,
-    config::Config,
-    env::Env,
-    process::ProcessBuilder,
-    term,
-};
+use crate::{cargo::Workspace, cli::Args, env::Env, process::ProcessBuilder, term};
 
 pub(crate) struct Context {
-    pub(crate) args: Args,
     pub(crate) env: Env,
-    pub(crate) config: Config,
-    pub(crate) verbose: bool,
-    pub(crate) target_dir: Utf8PathBuf,
-    pub(crate) doctests_dir: Utf8PathBuf,
-    pub(crate) package_name: String,
-    pub(crate) profdata_file: Utf8PathBuf,
+    pub(crate) ws: Workspace,
 
-    // cargo workspace info
-    pub(crate) metadata: cargo_metadata::Metadata,
-    // package root
-    pub(crate) manifest_path: Utf8PathBuf,
+    pub(crate) args: Args,
+    pub(crate) verbose: bool,
+
     pub(crate) workspace_members: WorkspaceMembers,
 
     // Paths to executables.
-    cargo: Cargo,
     pub(crate) llvm_cov: Utf8PathBuf,
     pub(crate) llvm_profdata: Utf8PathBuf,
 }
@@ -38,21 +23,10 @@ pub(crate) struct Context {
 impl Context {
     pub(crate) fn new(mut args: Args) -> Result<Self> {
         let mut env = Env::new()?;
-
-        let package_root = cargo::package_root(&env, args.manifest_path.as_deref())?;
-        let metadata = cargo::metadata(&env, &package_root)?;
-
-        let cargo = Cargo::new(&env, &metadata.workspace_root)?;
-
-        let config = Config::new(&cargo, &metadata.workspace_root)?;
-        config.merge_to(&mut args, &mut env);
-
+        let ws = Workspace::new(&env, args.manifest_path.as_deref())?;
+        ws.config.merge_to_env(&mut env);
+        ws.config.merge_to_args(&mut args.target, &mut args.verbose, &mut args.color);
         term::set_coloring(&mut args.color);
-
-        if let Some(v) = env::var_os("LLVM_PROFILE_FILE") {
-            warn!("environment variable LLVM_PROFILE_FILE={:?} will be ignored", v);
-            env::remove_var("LLVM_PROFILE_FILE");
-        }
 
         args.html |= args.open;
         if args.output_dir.is_some() && !args.show() {
@@ -82,19 +56,14 @@ impl Context {
             );
         }
         if args.output_dir.is_none() && args.html {
-            args.output_dir = Some(metadata.target_directory.join("llvm-cov"));
+            args.output_dir = Some(ws.output_dir.clone());
         }
-
-        // If we change RUSTFLAGS, all dependencies will be recompiled. Therefore,
-        // use a subdirectory of the target directory as the actual target directory.
-        let target_dir = metadata.target_directory.join("llvm-cov-target");
-        let doctests_dir = target_dir.join("doctestbins");
 
         // target-libdir (without --target flag) returns $sysroot/lib/rustlib/$host_triple/lib
         // llvm-tools exists in $sysroot/lib/rustlib/$host_triple/bin
         // https://github.com/rust-lang/rust/issues/85658
         // https://github.com/rust-lang/rust/blob/595088d602049d821bf9a217f2d79aea40715208/src/bootstrap/dist.rs#L2009
-        let mut rustlib: Utf8PathBuf = cargo.rustc_print("target-libdir")?.into();
+        let mut rustlib: Utf8PathBuf = ws.rustc_print("target-libdir")?.into();
         rustlib.pop(); // lib
         rustlib.push("bin");
         let llvm_cov = rustlib.join(format!("{}{}", "llvm-cov", env::consts::EXE_SUFFIX));
@@ -104,41 +73,22 @@ impl Context {
         if !llvm_cov.exists() || !llvm_profdata.exists() {
             bail!(
                 "failed to find llvm-tools-preview, please install llvm-tools-preview with `rustup component add llvm-tools-preview{}`",
-                if cargo.nightly { "" } else { " --toolchain nightly" }
+                if ws.cargo.nightly { "" } else { " --toolchain nightly" }
             );
         }
 
-        let package_name = metadata.workspace_root.file_stem().unwrap().to_string();
-        let profdata_file = target_dir.join(format!("{}.profdata", package_name));
-
-        let workspace_members = WorkspaceMembers::new(&args, &metadata);
+        let workspace_members = WorkspaceMembers::new(&args, &ws.metadata);
         if workspace_members.included.is_empty() {
             bail!("no crates to be measured for coverage");
         }
 
         let verbose = args.verbose != 0;
-        let manifest_path = package_root;
-        Ok(Self {
-            args,
-            env,
-            config,
-            verbose,
-            target_dir,
-            doctests_dir,
-            package_name,
-            profdata_file,
-            metadata,
-            manifest_path,
-            workspace_members,
-            cargo,
-            llvm_cov,
-            llvm_profdata,
-        })
+        Ok(Self { env, ws, args, verbose, workspace_members, llvm_cov, llvm_profdata })
     }
 
     pub(crate) fn process(&self, program: impl Into<OsString>) -> ProcessBuilder {
         let mut cmd = cmd!(program);
-        cmd.dir(&self.metadata.workspace_root);
+        cmd.dir(&self.ws.metadata.workspace_root);
         // cargo displays env vars only with -vv.
         if self.args.verbose > 1 {
             cmd.display_env_vars();
@@ -147,21 +97,7 @@ impl Context {
     }
 
     pub(crate) fn cargo_process(&self) -> ProcessBuilder {
-        let mut cmd = self.cargo.process();
-        cmd.dir(&self.metadata.workspace_root);
-        // cargo displays env vars only with -vv.
-        if self.args.verbose > 1 {
-            cmd.display_env_vars();
-        }
-        cmd
-    }
-}
-
-impl ops::Deref for Context {
-    type Target = Args;
-
-    fn deref(&self) -> &Self::Target {
-        &self.args
+        self.ws.cargo_process(self.args.verbose)
     }
 }