+++ src/fs.rs
@@ -1,6 +1,7 @@
-use std::{io, path::Path};
+pub(crate) use std::fs::Metadata;
+use std::{ffi::OsStr, io, path::Path};
 
-pub(crate) use fs_err::{create_dir_all, read_dir, write};
+pub(crate) use fs_err::{create_dir_all, read_dir, symlink_metadata, write};
 
 /// Removes a file from the filesystem **if exists**.
 pub(crate) fn remove_file(path: impl AsRef<Path>) -> io::Result<()> {
@@ -16,4 +17,15 @@ pub(crate) fn remove_dir_all(path: impl AsRef<Path>) -> io::Result<()> {
         Err(e) if e.kind() == io::ErrorKind::NotFound => Ok(()),
         res => res,
     }
+}
+
+pub(crate) fn file_stem_recursive(path: &Path) -> Option<&OsStr> {
+    let mut file_name = path.file_name()?;
+    while let Some(stem) = Path::new(file_name).file_stem() {
+        if file_name == stem {
+            break;
+        }
+        file_name = stem;
+    }
+    Some(file_name)
 }